requirements.txt
streamlit
google.generativeai
langchain